requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
reportlab>=4.0.0
kaleido>=0.2.1
PyYAML>=6.0
